Rename ECE components
